+++ changelog.md
@@ -4,6 +4,8 @@
 * Night Vision Fade out time is now configurable
 * Brightness forcing is now more configurable
 * Changed red outline on wrong tool to no outline on wrong tool
+* Combat Knockback feature has been cleaned up
+  * No weapon penalty and spam penalty are now two separate config options
 * Fast foods min speed to eat has been reduced (0.75s -> 0.6s)
 * Smokers now require soul sand/soil to craft
 * Changed, moved and renamed many config options